refactor: extract background loop mixin from engine/task_engine.py
