+++ src/synthorg/engine/task_engine.py
@@ -8,7 +8,6 @@ Observer notifications are dispatched via a separate background queue.
 
 import asyncio
 import contextlib
-from dataclasses import dataclass, field
 from typing import TYPE_CHECKING, Never
 from uuid import uuid4
 
@@ -22,11 +21,10 @@ from synthorg.engine.errors import (
     TaskNotFoundError,
     TaskVersionConflictError,
 )
-from synthorg.engine.task_engine_apply import dispatch as _dispatch_mutation
 from synthorg.engine.task_engine_config import TaskEngineConfig
-from synthorg.engine.task_engine_events import (
-    build_state_changed_event,
-    publish_snapshot,
+from synthorg.engine.task_engine_loops import (
+    TaskEngineLoopsMixin,
+    _MutationEnvelope,
 )
 from synthorg.engine.task_engine_models import (
     CancelTaskMutation,
@@ -43,19 +41,9 @@ from synthorg.engine.task_engine_version import VersionTracker
 from synthorg.observability import get_logger
 from synthorg.observability.events.task_engine import (
     TASK_ENGINE_CREATED,
-    TASK_ENGINE_DRAIN_COMPLETE,
-    TASK_ENGINE_DRAIN_START,
-    TASK_ENGINE_DRAIN_TIMEOUT,
-    TASK_ENGINE_FUTURES_FAILED,
     TASK_ENGINE_LIST_CAPPED,
-    TASK_ENGINE_LOOP_ERROR,
-    TASK_ENGINE_MUTATION_APPLIED,
     TASK_ENGINE_MUTATION_FAILED,
-    TASK_ENGINE_MUTATION_RECEIVED,
     TASK_ENGINE_NOT_RUNNING,
-    TASK_ENGINE_OBSERVER_DRAIN_TIMEOUT,
-    TASK_ENGINE_OBSERVER_FAILED,
-    TASK_ENGINE_OBSERVER_QUEUE_FULL,
     TASK_ENGINE_QUEUE_FULL,
     TASK_ENGINE_READ_FAILED,
     TASK_ENGINE_STARTED,
@@ -73,21 +61,7 @@ if TYPE_CHECKING:
 logger = get_logger(__name__)
 
 
-@dataclass
-class _MutationEnvelope:
-    """Pairs a mutation request with its response future.
-
-    Note: must be instantiated within a running event loop (the
-    ``future`` default factory calls ``asyncio.get_running_loop()``).
-    """
-
-    mutation: TaskMutation
-    future: asyncio.Future[TaskMutationResult] = field(
-        default_factory=lambda: asyncio.get_running_loop().create_future(),
-    )
-
-
-class TaskEngine:
+class TaskEngine(TaskEngineLoopsMixin):
     """Centralized single-writer for all task state mutations.
 
     Actor-like pattern: mutations are queued, processed sequentially,
@@ -200,92 +174,6 @@ class TaskEngine:
         await self._drain_observer(observer_budget)
 
         logger.info(TASK_ENGINE_STOPPED)
-
-    async def _drain_processing(self, budget: float) -> None:
-        """Drain the mutation processing loop within *budget* seconds."""
-        if self._processing_task is None:
-            return
-        logger.info(
-            TASK_ENGINE_DRAIN_START,
-            pending=self._queue.qsize(),
-            timeout_seconds=budget,
-        )
-        try:
-            await asyncio.wait_for(
-                asyncio.shield(self._processing_task),
-                timeout=budget,
-            )
-            logger.info(TASK_ENGINE_DRAIN_COMPLETE)
-        except TimeoutError:
-            logger.warning(
-                TASK_ENGINE_DRAIN_TIMEOUT,
-                remaining=self._queue.qsize(),
-            )
-            saved_in_flight = self._in_flight
-            self._processing_task.cancel()
-            with contextlib.suppress(asyncio.CancelledError):
-                await self._processing_task
-            self._fail_remaining_futures(saved_in_flight)
-        except BaseException:
-            self._fail_remaining_futures(self._in_flight)
-            raise
-        finally:
-            self._processing_task = None
-
-    async def _drain_observer(self, budget: float) -> None:
-        """Drain the observer dispatch loop within *budget* seconds."""
-        if self._observer_task is None:
-            return
-        try:
-            await asyncio.wait_for(
-                asyncio.shield(self._observer_task),
-                timeout=budget,
-            )
-        except TimeoutError:
-            logger.warning(
-                TASK_ENGINE_OBSERVER_DRAIN_TIMEOUT,
-                remaining=self._observer_queue.qsize(),
-            )
-            self._observer_task.cancel()
-            with contextlib.suppress(asyncio.CancelledError):
-                await self._observer_task
-        finally:
-            self._observer_task = None
-
-    def _fail_remaining_futures(
-        self,
-        saved_in_flight: _MutationEnvelope | None = None,
-    ) -> None:
-        """Fail in-flight and queued futures after drain timeout."""
-        shutdown_result_for = self._shutdown_result
-        failed_count = 0
-        in_flight = saved_in_flight if saved_in_flight is not None else self._in_flight
-        if in_flight is not None and not in_flight.future.done():
-            in_flight.future.set_result(shutdown_result_for(in_flight))
-            failed_count += 1
-        self._in_flight = None
-        while not self._queue.empty():
-            with contextlib.suppress(asyncio.QueueEmpty):
-                envelope = self._queue.get_nowait()
-                if not envelope.future.done():
-                    envelope.future.set_result(shutdown_result_for(envelope))
-                    failed_count += 1
-        if failed_count:
-            logger.warning(
-                TASK_ENGINE_FUTURES_FAILED,
-                failed_futures=failed_count,
-                note="Resolved remaining futures with shutdown failure",
-            )
-
-    @staticmethod
-    def _shutdown_result(envelope: _MutationEnvelope) -> TaskMutationResult:
-        """Build an internal-failure result for a shutdown-aborted envelope."""
-        return TaskMutationResult(
-            request_id=envelope.mutation.request_id,
-            success=False,
-            error="TaskEngine shut down before processing",
-            error_code="internal",
-        )
 
     @property
     def is_running(self) -> bool:
@@ -651,171 +539,3 @@ class TaskEngine:
 
     _SNAPSHOT_CHANNEL: str = "tasks"
     """Snapshot channel (must match ``CHANNEL_TASKS`` in ``api.channels``)."""
-
-    async def _processing_loop(self) -> None:
-        """Background loop: dequeue and process mutations sequentially.
-
-        Continues draining queued mutations after ``_running`` is set to
-        ``False``, enabling graceful shutdown.
-        """
-        while self._running or not self._queue.empty():
-            try:
-                envelope = await asyncio.wait_for(
-                    self._queue.get(),
-                    timeout=self._POLL_INTERVAL_SECONDS,
-                )
-            except TimeoutError:
-                continue
-            try:
-                await self._process_one(envelope)
-            except (MemoryError, RecursionError) as exc:
-                if not envelope.future.done():
-                    envelope.future.set_exception(exc)
-                raise
-            except Exception:
-                logger.exception(
-                    TASK_ENGINE_LOOP_ERROR,
-                    error="Unhandled exception in processing loop",
-                )
-                if not envelope.future.done():
-                    envelope.future.set_result(
-                        TaskMutationResult(
-                            request_id=envelope.mutation.request_id,
-                            success=False,
-                            error="Internal error in processing loop",
-                            error_code="internal",
-                        ),
-                    )
-
-    async def _process_one(self, envelope: _MutationEnvelope) -> None:
-        """Process a single mutation envelope."""
-        mutation = envelope.mutation
-        self._in_flight = envelope
-        logger.debug(
-            TASK_ENGINE_MUTATION_RECEIVED,
-            mutation_type=mutation.mutation_type,
-            request_id=mutation.request_id,
-        )
-        try:
-            result = await _dispatch_mutation(
-                mutation,
-                self._persistence,
-                self._versions,
-            )
-            if not envelope.future.done():
-                envelope.future.set_result(result)
-            if result.success:
-                task_id = getattr(mutation, "task_id", None)
-                logger.info(
-                    TASK_ENGINE_MUTATION_APPLIED,
-                    mutation_type=mutation.mutation_type,
-                    request_id=mutation.request_id,
-                    task_id=task_id or (result.task.id if result.task else None),
-                    version=result.version,
-                    previous_status=(
-                        result.previous_status.value if result.previous_status else None
-                    ),
-                    new_status=(result.task.status.value if result.task else None),
-                )
-            if result.success and (self._config.publish_snapshots or self._observers):
-                event = build_state_changed_event(mutation, result)
-                if self._config.publish_snapshots:
-                    await publish_snapshot(
-                        mutation,
-                        event,
-                        message_bus=self._message_bus,
-                        sender=self._SNAPSHOT_SENDER,
-                        channel=self._SNAPSHOT_CHANNEL,
-                    )
-                if self._observers:
-                    try:
-                        self._observer_queue.put_nowait(event)
-                    except asyncio.QueueFull:
-                        logger.warning(
-                            TASK_ENGINE_OBSERVER_QUEUE_FULL,
-                            request_id=event.request_id,
-                            task_id=event.task_id,
-                            mutation_type=event.mutation_type,
-                            queue_size=self._observer_queue.qsize(),
-                        )
-        except MemoryError, RecursionError:
-            raise
-        except Exception as exc:
-            internal_msg = f"{type(exc).__name__}: {exc}"
-            logger.exception(
-                TASK_ENGINE_MUTATION_FAILED,
-                mutation_type=mutation.mutation_type,
-                request_id=mutation.request_id,
-                error=internal_msg,
-            )
-            if not envelope.future.done():
-                envelope.future.set_result(
-                    TaskMutationResult(
-                        request_id=mutation.request_id,
-                        success=False,
-                        error="Internal error processing mutation",
-                        error_code="internal",
-                    ),
-                )
-        finally:
-            self._in_flight = None
-
-    # -- Event building & notification ------------------------------------
-
-    async def _observer_dispatch_loop(self) -> None:
-        """Background loop: dequeue and dispatch observer events.
-
-        Mirrors ``_processing_loop`` but for observer notifications.
-        Exits when a ``None`` sentinel is dequeued (placed by
-        ``stop()`` after the processing loop finishes).
-        """
-        while True:
-            try:
-                event = await asyncio.wait_for(
-                    self._observer_queue.get(),
-                    timeout=self._POLL_INTERVAL_SECONDS,
-                )
-            except TimeoutError:
-                continue
-            if event is None:
-                self._observer_queue.task_done()
-                break  # sentinel -- processing loop is done
-            try:
-                await self._notify_observers(event)
-            except MemoryError, RecursionError:
-                raise
-            except Exception:
-                logger.exception(
-                    TASK_ENGINE_LOOP_ERROR,
-                    error="Unhandled exception in observer dispatch loop",
-                    task_id=event.task_id,
-                    request_id=event.request_id,
-                    mutation_type=event.mutation_type,
-                )
-            finally:
-                self._observer_queue.task_done()
-
-    async def _notify_observers(
-        self,
-        event: TaskStateChanged,
-    ) -> None:
-        """Notify registered observers of a successful mutation.
-
-        Best-effort: observer errors are logged and swallowed
-        (except ``MemoryError`` and ``RecursionError``, which
-        propagate).
-        """
-        for observer in self._observers:
-            try:
-                await observer(event)
-            except MemoryError, RecursionError:
-                raise
-            except Exception:
-                logger.warning(
-                    TASK_ENGINE_OBSERVER_FAILED,
-                    observer=getattr(observer, "__qualname__", repr(observer)),
-                    mutation_type=event.mutation_type,
-                    request_id=event.request_id,
-                    task_id=event.task_id,
-                    exc_info=True,
-                )
